Viewer: implement pan with right and middle mouse button
Change panning behavior such that it is possible to pan the map with the
right and middle mouse button. This is useful for the other mouse modes
(selection tools). To make pan behavior better controlable
by users of Viewer and AbstractViewer, add methods setPanWithButton1()
and setZoomByDoubleClick() to fine tune what events are being processed
by the viewer.