Create new modules config, tools and places
This reduces the size of module gis and makes it easier for third-party
projects to use parts of the jeography codebase without pulling in too
much of the heavy gis module. Also this step avoids a circular
dependency among jeography-gis and osmocrat-core that we would otherwise
have in the future (because module gis is using osmocrat-core as a
dependency since recently, and osmocrat in turn depends on
jeography-gis). With the new modularization, we can make osmocrat depend
on jeography-tools instead.